vacation: error if only --from or --to is given
I think we could allow a --from (and implicetely assume today as the
--to), but the way it currently works is a bit weird if --from is
omitted (as we then use the inception day, and would add thousands of
vacation days).

For now, maybe we should stick to the simple fix of allowing only --from
and --to in combination (at least for vacation days).