add (preliminary) support for ignored days
I'm not too happy about the duplication of code and ideas between
vacation days and ignored days, but I kinda needed the feature myself so
this is a hacky, hopefully temporary solution.

Ideally, maybe we can combine all sorts of ignored days in one command
(watson-worktime ignore --vacation ..., watson-worktime ignore --sick
...). The aliases `watson-worktime vacation add` could stay. We can then
define rules which ignored days are still counted in overtime (overtime
during vacation? count it! overtime during ignored days? ignore it!).

In addition, we can combine the format to use a single file only, and we
can make the different ignore "modes" exclusive (a day is either a
vacation day or a sick day, not both at the same time).

Fractional vacation/sick days might be a bit hard then, though. We'll
see.

Anyway, for now it works \o/